+++ ChangeLog.md
@@ -1,3 +1,7 @@
+# 0.20
+- BREAKING CHANGE ValueChangedEventArgs swap arguments and rename PreviousValue to OldValue, so that you get a compiler error on event handlers. If you create ValueChangedEventArgs you have to swap the argument manually. No Warning in this case!
+- add SimpleCommand.RaiseCanExecuteChanged
+- 
 # 0.19
 - convert to SDK format
 - TargetFrameworks: net5.0-windows, netcoreapp3.1, net45